add StorageDead
