+++ nll-repr/src/repr/parser.lalrpop
@@ -99,6 +99,7 @@ Action: Action = {
     <c:Constraint> ";" Comment? => Action::Constraint(c),
     "use" "(" <v:Path> ")" ";" Comment? => Action::Use(v),
     "drop" "(" <v:Path> ")" ";" Comment? => Action::Drop(v),
+    "StorageDead" "(" <v:Variable> ")" ";" Comment? => Action::StorageDead(v),
     ";" Comment? => Action::Noop,
 };
