+++ nll/src/liveness.rs
@@ -1,4 +1,3 @@
-use def_use::DefUse;
 use env::{Environment, Point};
 use graph::{BasicBlockIndex, FuncGraph};
 use graph_algorithms::Graph;
@@ -172,3 +171,36 @@ impl Liveness {
     }
 }
 
+pub trait DefUse {
+    /// Returns (defs, uses), where `defs` contains variables whose
+    /// current value is completely overwritten, and `uses` contains
+    /// variables whose current value is used. Note that a variable
+    /// may exist in both sets.
+    fn def_use(&self) -> (Vec<repr::Variable>, Vec<repr::Variable>);
+}
+
+impl DefUse for repr::Action {
+    fn def_use(&self) -> (Vec<repr::Variable>, Vec<repr::Variable>) {
+        match *self {
+            repr::Action::Borrow(ref p, _name, _, ref q) => (vec![p.base()], vec![q.base()]),
+            repr::Action::Init(ref a, ref params) => {
+                (a.write_def().into_iter().collect(),
+                 params.iter().map(|p| p.base()).chain(a.write_use()).collect())
+            }
+            repr::Action::Assign(ref a, ref b) => {
+                (a.write_def().into_iter().collect(),
+                 once(b.base()).chain(a.write_use()).collect())
+            }
+            repr::Action::Constraint(ref _c) => (vec![], vec![]),
+            repr::Action::Use(ref v) => (vec![], vec![v.base()]),
+
+            // drop is special; it is not considered a "full use" of
+            // the variable that is being dropped
+            repr::Action::Drop(..) => (vec![], vec![]),
+
+            repr::Action::Noop => (vec![], vec![]),
+
+            repr::Action::StorageDead(_) => (vec![], vec![]),
+        }
+    }
+}
