+++ nll/src/main.rs
@@ -18,7 +18,6 @@ use std::io::Read;
 use std::process;
 
 #[macro_use] mod log;
-mod def_use;
 mod env;
 use self::env::Environment;
 mod infer;
